spell modification, extern configuration, some minor fixes

+++ src/Spell.cpp
@@ -4,7 +4,7 @@ using namespace Polycode;
 
 const String Spell::CAST_SPELL = "cast_spell";
 
-Spell::Spell( const Polycode::String& filename, int width, int height )
+Spell::Spell( const Polycode::String& filename, uint16 width, uint16 height )
 	:	ScreenSprite(filename, width, height)
 {
